\name{gbRd-package}
\alias{gbRd-package}
\alias{gbRd}
\docType{package}
\title{
Utilities for processing Rd objects and files
}
% Provides utilities for processing Rd objects and files.
% Extract argument descriptions and other parts of the help pages of
% functions.
%
% In the version of gbRd on 2011-10-31 (and 2.13 of R), if the
% description consists of the above lines, the `Details' header happens
% towards the top of the page but everything after it goes on the
% following page (leaving an almost empty page.
%

\description{
Provides utilities for processing Rd objects and files.
Extract argument descriptions and other parts of the help pages of
functions. The main purpose of the functions is to facilitate
extraction of descriptions of function arguments for presentation
of simplified usage descriptions.
}

% \details{
% 
% }

\author{
Georgi N. Boshnakov

Maintainer: Georgi N. Boshnakov <georgi.boshnakov@manchester.ac.uk>
}
\keyword{ package }

