\name{gbRd-package}
\alias{gbRd-package}
\alias{gbRd}
\docType{package}
\title{
Utilities for processing Rd objects and files
}

\description{
Provides utilities for processing Rd objects and files.
Extract argument descriptions and other parts of the help pages of
functions. The main purpose of the functions is to facilitate
extraction of descriptions of function arguments for presentation
of simplified usage descriptions.
}

\author{
Georgi N. Boshnakov

Maintainer: Georgi N. Boshnakov <georgi.boshnakov@manchester.ac.uk>
}
\keyword{ package }
