\name{gbRd-package}
\alias{gbRd-package}
\alias{gbRd}
\docType{package}
\title{
Utilities for processing Rd objects and files
}
\description{
Provides utilities for processing Rd objects and files.
}
\details{
\tabular{ll}{
Package: \tab gbRd\cr
Type: \tab Package\cr
Version: \tab 0.03\cr
Date: \tab 2011-01-09\cr
License: \tab GPL2 or later\cr
LazyLoad: \tab yes\cr
Built: \tab R 2.12.1; ; 2011-01-09 20:37:44 UTC; windows\cr
}

Index:
\preformatted{
Rd_fun                  Return all or selected fields from a help topic
                        as an Rd object
Rd_help2txt             Extract selected help sections as text.
Rd_title                Wrap an object so that it can be used as a
                        section element of an Rd object.
Rdo_args2txt            Extract the descriptions of one or more
                        arguments of a function
Rdo_args2txt_list       Extract the descriptions of the arguments of a
                        function
Rdo_create              Create a basic Rd object
Rdo_empty               Create an empty Rd object
Rdo_get_args            Extract argument description from a help topic
Rdo_section             Extract a section element from an Rd object or
                        Rd fragment
gbRd-package            Utilities for Rd objects
}
}
\author{
Georgi N. Boshnakov

Maintainer: Georgi N. Boshnakov <georgi.boshnakov@manchester.ac.uk>
}
\keyword{ package }
